refactor: move `coverage.ts` to `coverage/c8.ts`
