+++ packages/vite-node/src/client.ts
@@ -216,7 +216,7 @@ export class ViteNodeRunner {
     // Be careful when changing this
     // changing context will change amount of code added on line :114 (vm.runInThisContext)
     // this messes up sourcemaps for coverage
-    // adjust `offset` variable in packages/vitest/src/integrations/coverage.ts#L100 if you do change this
+    // adjust `offset` variable in packages/vitest/src/integrations/coverage/c8.ts#L100 if you do change this
     const context = this.prepareContext({
       // esm transformed by Vite
       __vite_ssr_import__: request,
